apps/api/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
requests
beautifulsoup4
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv
cachetools
